feat: Implement onTermination() for process exit signal handling (#12)
* feat: Implement onTermination() for process exit signal handling

The new function onTermination() can be used to register callbacks for
when the process receives a SIGTERM or SIGINT signal. These callbacks
will be run in reverse (compared to the order in which they were added).
After processing all callbacks, the process will exit explicitly.
Callbacks can be async by returning a Promise.

Tests are added for this behavior on POSIX systems. These same tests
cannot be replicated on Windows due to lack of signal support. Still,
Node.js will emulate SIGINT when CTRL+C is pressed in a Windows shell,
meaning the effectiveness of this code on Windows is still pretty much
guaranteed, we simply cannot run the tests.

* tests: Fix line terminator handling in termination.test.ts